加入Nacos扩展配置(mybatis-plus-dev.yaml) 添加自定义异常类(MyException.class) 新建gulimall-feign远程调用模块 ---商品满减信息SkuFullReductionController(保存sku的优惠、满减等信息)接口--- ---商品spu积分设置SpuBoundsController(保存spu积分信息)接口--- ---属性分组Controller(根据分类id获取分类下所有分组)接口--- ---spu信息SpuInfoController(根据分类id获取分类下所有分组)接口--- 添加(SkuReductionTo--SpuBoundTo--SpuSaveVo)

+++ gulimall-product/src/main/java/com/hngc/product/entity/ProductAttrValue.java
@@ -46,5 +46,5 @@ public class ProductAttrValue implements Serializable {
     private Integer attrSort;
 
     @ApiModelProperty("快速展示【是否展示在介绍上；0-否 1-是】")
-    private Byte quickShow;
+    private Integer quickShow;
 }
